Move icon to separate file, use for Account Security page as well
