+++ client/web/src/components/externalAccounts/externalAccounts.ts
@@ -6,6 +6,7 @@ import GithubIcon from 'mdi-react/GithubIcon'
 import GitLabIcon from 'mdi-react/GitlabIcon'
 
 import { AuthProvider } from '../../jscontext'
+import GerritIcon from '../externalServices/GerritIcon'
 
 export type ExternalAccountKind = Exclude<
     AuthProvider['serviceType'],
@@ -47,6 +48,6 @@ export const defaultExternalAccounts: Record<ExternalAccountKind, ExternalAccoun
     },
     gerrit: {
         title: 'Gerrit',
-        icon: AccountCircleIcon,
+        icon: GerritIcon,
     },
 }
